Always pass all datagram contents to Connection in one go
